### 2025-12-07 #9569 @Yehor-Mishchyriak
Hello.

Quick update: When I submitted the paper, I noted that no external funding supported the work. I realized that, because this project was done under Professor Kelly M. Thayer’s supervision at Wesleyan (individual tutorial / lab work), it was in fact supported by the National Science Foundation under Grant No. CHE-2320718. I have updated the paper's acknowledgements section to reflect this. Apologies for the oversight. Happy to provide anything else you need.
Unrelated: I added citations for micrograd, and numpy-ml.

Thank you for your time!
