### 2025-12-07 be3d084
public: release 1.2.7
