+++ JOSS/paper.md
@@ -1,0 +1,136 @@
+---
+title: "PureML: a transparent NumPy-only deep learning framework for teaching and prototyping"
+tags:
+  - python
+  - numpy
+  - autodiff
+  - machine-learning
+  - deep-learning
+  - education
+authors:
+  - name: Yehor Mishchyriak
+    orcid: 0009-0001-8371-7159
+    email: ymishchyriak@wesleyan.edu
+    affiliation: 1
+affiliations:
+  - name: Department of Mathematics and Computer Science at Wesleyan University, Middletown, CT, United States
+    index: 1
+date: 1 December 2025
+bibliography: paper.bib
+---
+
+# Summary
+PureML is a compact deep-learning framework implemented entirely in NumPy. It provides a tensor type with reverse-mode automatic differentiation, core neural-network layers and losses, optimizers and learning-rate schedulers, activations, training utilities, and persistence for model and optimizer states. A packaged MNIST dataset makes it easy to benchmark or teach end-to-end. The package targets learners, educators, and researchers who want to inspect and prototype deep-learning ideas with full control and transparency over the code involved, without heavy multi-language codebases that primarily optimize for production deployments and obscure the machine-learning theory underlying the implementation.
+
+# Statement of need
+Modern deep-learning libraries such as PyTorch and JAX provide rich ecosystems but are conceptually and operationally heavy for teaching low-level ML theory, code reading, or CPU-only environments [@paszke2019pytorch; @bradbury2018jax]. Educational materials often rely on pseudo-code or small snippets that omit practical details: broadcasting semantics, batching, parameter persistence, computational graph construction, vectorization, gradient accumulation, checkpointing, etc. As a result, learners struggle to bridge the gap to real systems. At the other end of the spectrum, educational projects like `micrograd` [@karpathy2020micrograd] purposefully keep the scope tiny (scalar or minimal tensor autodiff with a small MLP), which makes them excellent teaching tools but leaves out the aforementioned real-world challenges. Projects like `numpy-ml` [@bourgin2019numpyml] offer a broad catalog of algorithms implemented in NumPy but are not built around an autodiff engine and are not aimed at performance; they serve as references rather than frameworks for training deep networks.
+
+PureML aims to sit between these extremes: small enough to audit end-to-end, but feature-complete enough for nontrivial models. The code remains transparent while supporting batch-vectorized computation, dynamic computational graphs, forward pass caching, persistence, and related functionality. It focuses on:
+
+- Explicit reverse-mode autodiff with readable vector-Jacobian products for every operation, so gradient flow is inspectable.
+- Minimal runtime dependencies (NumPy and zarr) suitable for laptops, classrooms, and CPU-only servers [@harris2020numpy; @zarrpython2025].
+- Ready-to-run MNIST example to demonstrate end-to-end training without additional downloads [@lecun1998mnist].
+- Persistence utilities that round-trip models, optimizer slots, and data for reproducible exercises or small experiments.
+
+# Design and implementation
+
[…]
