+++ JOSS/paper.bib
@@ -1,0 +1,70 @@
+@article{harris2020numpy,
+  title = {Array programming with {NumPy}},
+  author = {Harris, Charles R and Millman, K Jarrod and van der Walt, St{\'e}fan J and others},
+  journal = {Nature},
+  volume = {585},
+  number = {7825},
+  pages = {357--362},
+  year = {2020},
+  doi = {10.1038/s41586-020-2649-2}
+}
+
+@software{zarrpython2025,
+  title = {zarr-developers/zarr-python: v3.1.5},
+  author = {Miles, Alistair and Kirkham, John and Stansby, David and Papadopoulos Orfanos, Dimitri and Hamman, Joe and others},
+  version = {v3.1.5},
+  year = {2025},
+  doi = {10.5281/zenodo.17672242},
+  url = {https://github.com/zarr-developers/zarr-python}
+}
+
+@article{paszke2019pytorch,
+  title = {PyTorch: An imperative style, high-performance deep learning library},
+  author = {Paszke, Adam and Gross, Sam and Massa, Francisco and others},
+  journal = {Advances in Neural Information Processing Systems},
+  volume = {32},
+  year = {2019},
+  url = {https://proceedings.neurips.cc/paper/2019/hash/bdbca288fee7f92f2bfa9f7012727740-Abstract.html}
+}
+
+@article{bradbury2018jax,
+  title = {JAX: composable transformations of {Python}+{NumPy} programs},
+  author = {Bradbury, James and Frostig, Roy and Hawkins, Peter and others},
+  year = {2018},
+  url = {https://github.com/google/jax}
+}
+
+@book{goodfellow2016deep,
+  title = {Deep Learning},
+  author = {Goodfellow, Ian and Bengio, Yoshua and Courville, Aaron},
+  publisher = {MIT Press},
+  year = {2016},
+  url = {https://www.deeplearningbook.org/}
+}
+
+@article{lecun1998mnist,
+  title = {Gradient-based learning applied to document recognition},
+  author = {LeCun, Yann and Bottou, Leon and Bengio, Yoshua and Haffner, Patrick},
+  journal = {Proceedings of the IEEE},
+  volume = {86},
+  number = {11},
+  pages = {2278--2324},
+  year = {1998},
+  doi = {10.1109/5.726791}
+}
+
+@software{karpathy2020micrograd,
+  title = {micrograd},
+  author = {Karpathy, Andrej},
+  year = {2020},
+  url = {https://github.com/karpathy/micrograd},
+  note = {MIT License}
+}
+
+@software{bourgin2019numpyml,
+  title = {numpy-ml},
+  author = {Bourgin, David},
+  year = {2019},
+  url = {https://github.com/ddbourgin/numpy-ml},
+  note = {GPL-3.0}
+}
