+++ tests/__init__.py
@@ -1,0 +1,17 @@
+from . import test_activations
+from . import test_knn
+from . import test_layers
+from . import test_machinery
+from . import test_optim_sched
+from . import test_storage
+from . import test_losses
+
+__all__ = [
+    "test_activations",
+    "test_knn",
+    "test_layers",
+    "test_machinery",
+    "test_optim_sched",
+    "test_storage",
+    "test_losses"
+]
