+++ JOSS/note_for_reviewers.md
@@ -1,0 +1,7 @@
+## Note for reviewers
+
+This public repository mirrors the private development repository in terms of source code, tests, and the submitted paper, but intentionally omits internal tooling and development infrastructure. For this reason, the full commit history is not reflected here. Day-to-day development occurs in a private workspace, and the public repository is kept minimal to avoid exposing development-specific files while providing reviewers and users with clean, direct access to the software.
+
+All code in both the private and public repositories was authored by me (Yehor Mishchyriak), and I am the sole contributor.
+
+If access to the full development history would be helpful during the review process, please let me know. Thank you for your time and consideration.

[…]
