### 2025-12-15 #9569 @espottesmith
Wow, this has been the snappiest and most overwhelmingly positive search for reviewers that I've been a part of. In the interest of limiting the number of metaphorical cooks in the kitchen, I'll stop things here and open the review.
