### 2025-12-15 #9631 @editorialbot
[REVIEW]: PureML: a transparent NumPy-only deep learning framework for teaching and prototyping
**Submitting author:** @Yehor-Mishchyriak (<a href="http://orcid.org/0009-0001-8371-7159">Yehor Mishchyriak</a>)
**Repository:** https://github.com/Yehor-Mishchyriak/PureML
**Branch with paper.md** (empty if default branch):
**Version:** v1.2.8

@sampottinger & @paquiteau & @Manishms18 & @yewentao256, your review will be checklist based. Each of you will have a separate checklist that you should update when carrying out your review.
