### 2025-12-15 Yehor-Mishchyriak/PureML#1 @paquiteau
[JOSS-Review] Packaging of PureML
Pure-ML works well, but its layout and setup does not follows the python packaging standards. I would recommend:

- [x] Use a pyproject.toml to declare the package metadata and dependencies
- [x] Move to a `src`-layout for the packaging (to avoid shipping test and examples with your package)
- [x] Add a Github CI for the release

Some references:
- https://packaging.python.org/en/latest/tutorials/packaging-projects/
- https://setuptools.pypa.io/en/latest/setuptools.html

JOSS-Review link: https://github.com/openjournals/joss-reviews/issues/9631
