### 2025-12-16 Yehor-Mishchyriak/PureML#1 @Yehor-Mishchyriak
Dear Pierre, thank you very much for your feedback!
Current model includes CI/CD on the private repo's end, which I intend to move on from and have a single repo with CI done thru github actions. Additionally, the overall project's structure does need some refinement, as you point out, to adhere to best practices. I will implement these changes this week -- most likely during the Friday thru Sunday period.

Thank you!
