### 2025-12-17 #9631 @yewentao256
Hi @Yehor-Mishchyriak , generally look good to me, several thoughts:

- Please add clear community guidelines (at least a CONTRIBUTING.md and a support/issue-reporting section, optionally a Code of Conduct).
- Please ensure the docs and README are consistent (e.g., `the seed argument is currently not applied` in the doc, but v1.2.7 — What’s new BUG FIX: DataLoader's seed)
- Since MNIST data is bundled/distributed with the package, please add a short statement about the dataset source and its licensing/terms of use.
- Please clarify the target audience and the concrete scenarios where PureML is materially better (e.g., simpler, more transparent, easier to audit/teach) than widely used alternatives, and strengthen the “state of the field” comparison with appropriate citations.
- Though not a requirement, I think it would be nice if you could provide some evidence of use (course material, tutorials, or downstream adoption) to support the claimed need and scope.
