### 2025-12-17 Yehor-Mishchyriak/PureML#2 @Yehor-Mishchyriak
[JOSS-Review] Documentation, dataset licensing, and scope clarification
> Hi @Yehor-Mishchyriak, several thoughts:
>
> - Please add clear community guidelines (at least a `CONTRIBUTING.md` and a support / issue-reporting section; optionally a Code of Conduct).
> - Please ensure the docs and README are consistent (e.g., the `seed` argument is currently not applied in the documentation, but is mentioned in v1.2.7 — *What’s new*: bug fix for `DataLoader` seed handling).
> - Since MNIST data is bundled/distributed with the package, please add a short statement about the dataset source and its licensing / terms of use.
> - Please clarify the target audience and the concrete scenarios where PureML is materially better (e.g., simpler, more transparent, easier to audit or teach) than widely used alternatives, and strengthen the “state of the field” comparison with appropriate citations.
> - Though not a requirement, it would be great to include some evidence of use (e.g., course material, tutorials, or downstream adoption) to support the claimed need and scope.
>
> — Comment by @yewentao256, reproduced here for reference

**Reference:**
https://github.com/openjournals/joss-reviews/issues/9631#issuecomment-3663141084
