### 2026-01-04 bcf3c04
Add pyproject packaging and CI release workflow
These are only initial changes. More refinement and testing will follow shortly
