### 2026-01-05 7cd62de
added pkg version to the root init
