+++ src/pureml/__init__.py
@@ -5,6 +5,8 @@ from .machinery import (
 )
 from .util import ArrayStorage, compose_steps, batches_of
 
+__version__ = "1.2.7"
+
 # expose subpackages on demand
 def __getattr__(name):
     if name in {
