+++ .github/workflows/release.yml
@@ -1,15 +1,15 @@
-name: CI & Release
+name: Release
 
 on:
   push:
-    branches: [main]
+    branches: [release]
   pull_request:
-    branches: [main]
-  release:
-    types: [published]
+    branches: [release]
+  workflow_dispatch: {}
 
 permissions:
   contents: read
+  id-token: write
 
 jobs:
   tests:
@@ -18,11 +18,8 @@ jobs:
       matrix:
         python-version: ["3.11", "3.12"]
     steps:
-      - name: Checkout repository
-        uses: actions/checkout@v4
-
-      - name: Set up Python
-        uses: actions/setup-python@v5
+      - uses: actions/checkout@v4
+      - uses: actions/setup-python@v5
         with:
           python-version: ${{ matrix.python-version }}
           cache: "pip"
@@ -37,31 +34,28 @@ jobs:
 
   publish:
     needs: tests
-    if: github.event_name == 'release' && github.event.action == 'published'
     runs-on: ubuntu-latest
+
     permissions:
       contents: read
-      id-token: write
-    steps:
-      - name: Checkout repository
-        uses: actions/checkout@v4
 
-      - name: Set up Python
-        uses: actions/setup-python@v5
+    steps:
+      - uses: actions/checkout@v4
+      - uses: actions/setup-python@v5
         with:
           python-version: "3.11"
           cache: "pip"
-
+  
       - name: Install build backend
         run: |
-          python -m pip install --upgrade pip
-          python -m pip install build twine
+          python3 -m pip install --upgrade pip
+          python3 -m pip install build twine
 
       - name: Build distributions
-        run: python -m build
+        run: python3 -m build
 
       - name: Verify distributions
-        run: python -m twine check dist/*
+        run: python3 -m twine check dist/*
 
       - name: Publish to PyPI
         uses: pypa/gh-action-pypi-publish@v1.10.1
