### 2026-01-05 e46df74
added separate testing workflow
