+++ .github/workflows/release.yml
@@ -1,8 +1,6 @@
 name: Release
 
 on:
-  push:
-    branches: [release]
   pull_request:
     branches: [release]
   workflow_dispatch: {}
