### 2026-01-05 Yehor-Mishchyriak/PureML#1 @Yehor-Mishchyriak
DONE!
