### 2026-01-05 fd084b6
added concise module level docstrings
