+++ src/pureml/datasets/MNIST/__init__.py
@@ -1,3 +1,4 @@
+"""MNIST dataset loader packaged with PureML."""
 from .dataset import MnistDataset
 
 __all__ = [
