+++ src/pureml/models/neural_networks/__init__.py
@@ -1,3 +1,4 @@
+"""Neural network reference models packaged with PureML."""
 from .mnist_beater import MNIST_BEATER
 
 __all__ = [
