+++ src/pureml/models/classical/__init__.py
@@ -1,3 +1,4 @@
+"""Classical ML models included with PureML."""
 from . import knn
 
 __all__ = [
