+++ src/pureml/models/__init__.py
@@ -1,1 +1,2 @@
+"""Packaged reference models (classical and neural network)."""
 __all__ = ["classical", "neural_networks"]
