### 2026-01-05 be44713
credited third-party libraries
