+++ CREDITS.MD
@@ -1,1 +1,5 @@
 At present, the only project maintainer and contributor is [Yehor Mishchyriak](https://github.com/Yehor-Mishchyriak).
+
+However, PureML builds on the excellent work of the following two communties: NumPy and Zarr:
+- [NumPy](https://numpy.org/) — core array library used throughout the framework.
+- [Zarr](https://zarr.readthedocs.io/) — storage backend for packaged datasets and checkpoints.
