### 2026-01-05 6d5e742
adjusted the workflow: now test -> build -> release only on v*-tagged pushes to main
