+++ .github/workflows/release.yml
@@ -1,9 +1,9 @@
 name: Release
 
 on:
-  pull_request:
-    branches: [release]
-  workflow_dispatch: {}
+  push:
+    tags:
+      - "v*"
 
 permissions:
   contents: read
@@ -11,12 +11,21 @@ permissions:
 
 jobs:
   tests:
+    if: startsWith(github.ref, 'refs/tags/v')
     runs-on: ubuntu-latest
     strategy:
       matrix:
         python-version: ["3.11", "3.12"]
     steps:
       - uses: actions/checkout@v4
+        with:
+          fetch-depth: 0
+
+      - name: Ensure tag commit is on main branch
+        run: |
+          git fetch --depth=1 origin main
+          git merge-base --is-ancestor "$GITHUB_SHA" origin/main
+
       - uses: actions/setup-python@v5
         with:
           python-version: ${{ matrix.python-version }}
@@ -31,6 +40,7 @@ jobs:
         run: python -m unittest discover -s tests
 
   publish:
+    if: startsWith(github.ref, 'refs/tags/v')
     needs: tests
     runs-on: ubuntu-latest
 
