### 2026-01-05 44f4cad
run tests upon push or PR to development branch or PR to main
