### 2026-01-05 fcf6ccd
made the module level docstrings more detailed
