+++ src/pureml/models/__init__.py
@@ -1,2 +1,2 @@
-"""Packaged reference models (classical and neural network)."""
+"""Packaged reference models: classical (kNN) and neural network (MNIST MLP)."""
 __all__ = ["classical", "neural_networks"]
