+++ src/pureml/models/classical/__init__.py
@@ -1,4 +1,4 @@
-"""Classical ML models included with PureML."""
+"""Classical ML models included with PureML (currently k-Nearest Neighbors)."""
 from . import knn
 
 __all__ = [
