+++ src/pureml/datasets/__init__.py
@@ -1,4 +1,4 @@
-"""Dataset wrappers packaged with PureML."""
+"""Dataset wrappers packaged with PureML (currently MNIST via bundled Zarr archive)."""
 from .MNIST import MnistDataset
 
 __all__ = [
