### 2026-01-05 Yehor-Mishchyriak/PureML#1 @Yehor-Mishchyriak
Thank you very much for the issue! Also apologies for closing it the first time, I will keep it open for now.
Please let me know (either here or at the JOSS issue) if any other changes are necessary.

Thanks!
