+++ docs.md
@@ -84,6 +84,7 @@ y = Tensor(5)                      # requires_grad=False by default
 - `.dtype`, `.shape`, `.ndim` mirror the underlying NumPy array.
 - `.numpy(copy=True, readonly=False)` returns a safe NumPy view/copy; prefer this over `.data` for read access.
 - `.detach()` returns a new leaf tensor sharing storage with `requires_grad=False`; `.detach_()` toggles in-place. `.requires_grad_(bool)` mirrors PyTorch.
+- Tensor constructor args: `data` (array-like or Tensor), `requires_grad=False`, `dtype=None` (casts input), `copy=False`, `ensure_writable=True` (copies if read-only), `coerce_float_if_grad=True` (ints/bools -> float when tracking grads).
 
 ### Gradient flow
 
@@ -116,11 +117,12 @@ x.zero_grad()         # sets grad to None
 
 ## Activations (`pureml.activations`)
 
-- `sigmoid(x)`
-- `relu(x)`
-- `tanh(x)`
-- `softmax(x, axis=-1)` - stable, axis-aware
-- `log_softmax(x, axis=-1)` - stable, axis-aware
+- `sigmoid(x: Tensor)`  
+- `relu(x: Tensor)`  
+- `tanh(x: Tensor)`  
+- `softmax(x: Tensor, axis=-1)` - stable, axis-aware  
+- `log_softmax(x: Tensor, axis=-1)` - stable, axis-aware  
+All accept any shape; `axis` is the class dimension for softmax/log_softmax. Return Tensors and provide Jacobian-free backward passes.
 
 Each returns a Tensor and has a Jacobian-free backward pass.
 
@@ -130,9 +132,14 @@ Each returns a Tensor and has a Jacobian-free backward pass.
 
 All losses return scalar tensors (mean over all elements/samples).
 
-- `MSE(Y, Y_hat)`: mean squared error.
-- `BCE(Y, Y_hat, from_logits=False, label_smoothing=0.0)`: binary cross-entropy on probabilities or logits (`from_logits=True`). Label smoothing moves targets toward 0.5.
-- `CCE(Y, Y_hat, from_logits=False, label_smoothing=0.0)`: categorical cross-entropy over the last axis. Accepts one-hot/soft labels. `from_logits=True` uses a stable log-softmax. Label smoothing mixes targets toward the uniform distribution over classes.
+- `MSE(Y, Y_hat)`  
+  - `Y`, `Y_hat`: broadcastable tensors. Returns mean squared error over all elements.
+- `BCE(Y, Y_hat, from_logits=False, label_smoothing=0.0)`  
+  - `Y`: targets in {0,1} or probabilities; `Y_hat`: probabilities or logits (set `from_logits=True`).  
+  - `label_smoothing` in [0,1): mixes targets toward 0.5. Returns mean over all elements.
+- `CCE(Y, Y_hat, from_logits=False, label_smoothing=0.0)`  
+  - `Y`: one-hot or soft labels; `Y_hat`: probabilities or logits (`from_logits=True`).  
+  - Operates over the last axis; label smoothing mixes targets toward uniform over classes. Returns mean over batch/classes.
 
 ---
 
@@ -141,16 +148,29 @@ All losses return scalar tensors (mean over all elements/samples).
 Common interface: `.parameters` (trainables), `.named_buffers()` (non-trainable state), `.train()/.eval()` toggle `training` and call `on_mode_change`.
 
 - **Affine(fan_in, fan_out, method="xavier-glorot-normal", W=None, b=None, bias=True, seed=None)**  
[…]
