### 2026-01-05 Yehor-Mishchyriak/PureML#2 @Yehor-Mishchyriak
Done! Thank you very much

Regarding "evidence of use", in the paper I mention [SAWNERGY](https://github.com/Yehor-Mishchyriak/SAWNERGY/blob/main/sawnergy/embedding/SGNS_pml.py)'s embedding pipeline that relies 100% on PureML.  The user can opt for torch there as well, but it's not a dependency while PureML is. SAWNERGY is applied at ThayerLab at Wesleyan University.
