### 2026-01-05 198a025
Fixed the typo: the tinygrad citation now credits George Hotz (not Hoffman)
