+++ JOSS/paper.bib
@@ -86,7 +86,7 @@
 
 @software{tinygrad2024,
   title = {tinygrad},
-  author = {Hoffman, George and contributors},
+  author = {Hotz, George and contributors},
   year = {2024},
   url = {https://github.com/tinygrad/tinygrad}
 }
