### 2026-01-17 4000557
Revert "added placeholders for future layers to implement"
This reverts commit 0add83763dca475ff7fb8ddfa7ff7cd9eb96b8b1.

Reverting the placeholders of future layers' addition because it's not
needed on main
