+++ src/pureml/layers.py
@@ -699,15 +699,6 @@ class BatchNorm1d(Layer):
         _logger.debug("BN1d.__call__: out.shape=%s", getattr(out.data, "shape", None))
         return out
 
-class LayerNorm1d(Layer):
-    """Layer Normalization for 2D inputs shaped (B, F).
-
-    Normalizes each sample across features:
-        y = gamma * (x - mu) / sqrt(var + eps) + beta
-    """
-
-    pass  # Implementation omitted for now
-
 class Embedding(Layer):
     """Learned lookup table: returns rows of `W` for integer indices.
 
@@ -935,141 +926,6 @@ class Embedding(Layer):
         
         return out
 
-class Conv1d(Layer):
-    """1D Convolutional layer.
-
-    Args:
-        in_channels: Number of input channels.
-        out_channels: Number of output channels.
-        kernel_size: Size of the convolutional kernel.
-        stride: Stride of the convolution.
-        padding: Padding added to both sides of input.
-        dilation: Spacing between kernel elements.
-        bias: Whether to include a bias term.
-        method: Initialization method for weights.
-        training: Initial mode flag.
-        seed: Optional RNG seed for initialization.
-    """
-
-    pass  # Implementation omitted for now
-
-class Conv2d(Layer):
-    """2D Convolutional layer.
-
-    Args:
-        in_channels: Number of input channels.
-        out_channels: Number of output channels.
-        kernel_size: Size of the convolutional kernel (height, width).
-        stride: Stride of the convolution (height, width).
-        padding: Padding added to both sides of input (height, width).
-        dilation: Spacing between kernel elements (height, width).
-        bias: Whether to include a bias term.
-        method: Initialization method for weights.
-        training: Initial mode flag.
-        seed: Optional RNG seed for initialization.
-    """
-
-    pass  # Implementation omitted for now
-
-class MaxPool1d(Layer):
-    """1D Max Pooling layer.
-
-    Args:
-        kernel_size: Size of the pooling kernel.
-        stride: Stride of the pooling operation.
-        padding: Padding added to both sides of input.
-    """
-
-    pass  # Implementation omitted for now
-
-class MaxPool2d(Layer):
-    """2D Max Pooling layer.
-
-    Args:
-        kernel_size: Size of the pooling kernel (height, width).
-        stride: Stride of the pooling operation (height, width).
-        padding: Padding added to both sides of input (height, width).
-    """
-
-    pass  # Implementation omitted for now
-
-class MeanPool1d(Layer):
-    """1D Mean Pooling layer.
-
-    Args:
-        kernel_size: Size of the pooling kernel.
-        stride: Stride of the pooling operation.
-        padding: Padding added to both sides of input.
-    """
-
-    pass  # Implementation omitted for now
-
-class MeanPool2d(Layer):
-    """2D Mean Pooling layer.
-
-    Args:
-        kernel_size: Size of the pooling kernel (height, width).
-        stride: Stride of the pooling operation (height, width).
-        padding: Padding added to both sides of input (height, width).
-    """
-
-    pass  # Implementation omitted for now
-
-class SumPool1d(Layer):
-    """1D Sum Pooling layer.
-
-    Args:
-        kernel_size: Size of the pooling kernel.
-        stride: Stride of the pooling operation.
-        padding: Padding added to both sides of input.
-    """
-
-    pass  # Implementation omitted for now
-
-class SumPool2d(Layer):
-    """2D Sum Pooling layer.
-
-    Args:
-        kernel_size: Size of the pooling kernel (height, width).
[…]
