### 2026-01-17 1697d6e
still trying to fix the DOI badge
