### 2026-01-17 bad5a95
finally fixed the doi badge
