### 2026-01-18 #9631 @espottesmith
@Yehor-Mishchyriak you should have your affiliation on the Zenodo, but otherwise, the archive looks good to me!
