### 2026-01-18 #9631 @espottesmith
The article seems to render correctly.

Some comments on the paper:
- There are a couple of places where you don't use complete sentences (e.g., the second paragraph of "Summary", all of "Availability"). Please use sentences throughout. Note that the bullet-list in "Statement of Need" is appropriate, because each bullet technically is a complete sentence when connected with the preceding "It focuses on".
- MNIST dataset should be cited at first mention (Summary)
- "Statement of Need" is well written and compelling. Well done.
-  "VJP" should be spelled out at first use; alternatively the initialism should be placed after the first use of the complete "vector-Jacobian products"
- Specific optimizers like Adam should be cited
