### 2026-01-19 #9631 @espottesmith
We are not currently looking for reviewers, no. We're at the post-review stage. But if you're looking to review for JOSS, please add yourself to our database (https://reviewers.joss.theoj.org/reviewers/).
