### 2026-01-20 aa2db73
Update JOSS citations and polish paper text
Update JOSS citations and polish paper text:
- Fix and standardize bib entries (Deep Learning, JAX, Keras)
- Add DOIs for PyTorch/TensorFlow and new optimizer/scheduler refs
- Cite MNIST at first mention; expand VJP on first use
- Convert Summary/Availability to full sentences
- Update optimizer section to cite SGD, AdaGrad, RMSProp, Adam, cosine annealing
