+++ JOSS/paper.bib
@@ -24,22 +24,69 @@
   journal = {Advances in Neural Information Processing Systems},
   volume = {32},
   year = {2019},
+  doi = {10.48550/arXiv.1912.01703},
   url = {https://proceedings.neurips.cc/paper/2019/hash/bdbca288fee7f92f2bfa9f7012727740-Abstract.html}
 }
 
-@article{bradbury2018jax,
-  title = {JAX: composable transformations of {Python}+{NumPy} programs},
-  author = {Bradbury, James and Frostig, Roy and Hawkins, Peter and others},
+@software{jax2018github,
+  author = {James Bradbury and Roy Frostig and Peter Hawkins and Matthew James Johnson and Chris Leary and Dougal Maclaurin and George Necula and Adam Paszke and Jake Vander{P}las and Skye Wanderman-{M}ilne and Qiao Zhang},
+  title = {{JAX}: composable transformations of {P}ython+{N}um{P}y programs},
+  url = {http://github.com/jax-ml/jax},
+  version = {0.3.13},
   year = {2018},
-  url = {https://github.com/google/jax}
 }
 
-@book{goodfellow2016deep,
-  title = {Deep Learning},
-  author = {Goodfellow, Ian and Bengio, Yoshua and Courville, Aaron},
-  publisher = {MIT Press},
+@book{Goodfellow-et-al-2016,
+    title={Deep Learning},
+    author={Ian Goodfellow and Yoshua Bengio and Aaron Courville},
+    publisher={MIT Press},
+    note={\url{http://www.deeplearningbook.org}},
+    year={2016}
+}
+
+@article{kingma2015adam,
+  title = {Adam: A method for stochastic optimization},
+  author = {Kingma, Diederik P and Ba, Jimmy},
+  year = {2015},
+  doi = {10.48550/arXiv.1412.6980},
+  url = {https://arxiv.org/abs/1412.6980}
+}
+
+@article{robbins1951stochastic,
+  title = {A stochastic approximation method},
+  author = {Robbins, Herbert and Monro, Sutton},
+  journal = {The Annals of Mathematical Statistics},
+  volume = {22},
+  number = {3},
+  pages = {400--407},
+  year = {1951},
+  doi = {10.1214/aoms/1177729586}
+}
+
+@article{duchi2011adagrad,
+  title = {Adaptive Subgradient Methods for Online Learning and Stochastic Optimization},
+  author = {Duchi, John and Hazan, Elad and Singer, Yoram},
+  journal = {Journal of Machine Learning Research},
+  volume = {12},
+  pages = {2121--2159},
+  year = {2011},
+  url = {http://jmlr.org/papers/v12/duchi11a.html}
+}
+
+@misc{tieleman2012rmsprop,
+  title = {Lecture 6.5 - rmsprop: Divide the gradient by a running average of its recent magnitude},
+  author = {Tieleman, Tijmen and Hinton, Geoffrey},
+  year = {2012},
+  howpublished = {Coursera: Neural Networks for Machine Learning},
+  url = {http://www.cs.toronto.edu/~tijmen/csc321/slides/lecture_slides_lec6.pdf}
+}
+
+@article{loshchilov2016sgdr,
+  title = {SGDR: Stochastic Gradient Descent with Warm Restarts},
+  author = {Loshchilov, Ilya and Hutter, Frank},
   year = {2016},
-  url = {https://www.deeplearningbook.org/}
+  doi = {10.48550/arXiv.1608.03983},
+  url = {https://arxiv.org/abs/1608.03983}
 }
 
 @article{lecun1998mnist,
@@ -74,14 +121,15 @@
[…]
