### 2026-01-20 #9631 @Yehor-Mishchyriak
Hello @espottesmith,
Thank you very much for the additional feedback! I have addressed all of the points you raised:
- Converted the Summary’s second paragraph and the Availability section to full sentences.
- Added the MNIST citation at its first mention in the Summary.
- Spelled out “vector-Jacobian products (VJPs)” at first use.
- Added citations for specific optimizers (Adam, SGD, AdaGrad, RMSProp) and cosine annealing.
- Added missing DOIs for PyTorch and TensorFlow, and included DOIs for the new citations wherever available.

Please let me know if any further changes are needed!
