### 2026-01-20 #9631 @espottesmith
Paper looks pretty good to me. Please add your affiliation to the Zenodo archive, and I'll be happy to recommend this for acceptance.
