### 2026-01-20 #9631 @Yehor-Mishchyriak
No worries at all, thank you for the clarification. When editing the record, Zenodo only suggested the institution in the affiliation dropdown, so I initially selected that to be safe. However, I have now updated the affiliation to include the department so that it matches the paper exactly.
