### 2026-01-23 #9631 @crvernon
🥳 Congratulations on your new publication @Yehor-Mishchyriak ! Many thanks to @espottesmith for editing and @sampottinger, @paquiteau  @Manishms18, and @yewentao256 for your time, hard work, and expertise!! JOSS wouldn't be able to function nor succeed without your efforts.
