### 2026-01-24 #9631 @Yehor-Mishchyriak
Thank you very much to everyone involved. I sincerely appreciate your time, effort, and the opportunity to publish this work with JOSS!
